refactor(harness): move command vocabulary from runtime
