+++ src/loushang/coding/ui/handlers.py
@@ -19,7 +19,7 @@ from loushang.coding.ui.lifecycle import RunLifecycle
 from loushang.coding.ui.pending_queue import pending_queue_view, restore_queued_messages
 from loushang.coding.ui.prompt_dispatch import PromptDispatchOutcome
 from loushang.coding.ui.prompt_routing import PromptRoute, route_prompt_intent
-from loushang.runtime.commands import CommandEffect, CommandEffectKind
+from loushang.harness.commands import CommandEffect, CommandEffectKind
 from loushang.tui import InfoPanel, PendingQueueView
